code cleanup
Move the controller related code to a new file
